+++ embassy-usb/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased - ReleaseDate
 
 - Bump usbd-hid from 0.9.0 to 0.10.0
+- `UAC1`: Add audio source
+- `UAC1`: `Speaker::new` now returns `Self` with the parts inside instead of a tuple
 
 ## 0.6.0 - 2026-03-10